requirement.txt
Django==2.1.4
django-filter==2.0.0
django-qr-code==1.0.0
djangorestframework==3.9.0
Pillow==5.3.0
pypng==0.0.18
PyQRCode==1.2.1
pytz==2018.7
qrcode==6.0
six==1.12.0
